# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для сервиса геолокации и точек интереса",
        "title": "AIGPS Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "45.150.8.131:8080",
    "basePath": "/",
    "paths": {
        "/api/poi/create": {
            "post": {
                "description": "Создает новую точку интереса с изображением и аудиофайлами",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "POI"
                ],
                "summary": "Создание новой точки интереса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название точки интереса",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание точки интереса",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Широта",
                        "name": "latitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Долгота",
                        "name": "longitude",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение точки интереса",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Короткое аудио",
                        "name": "short_audio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "Полные аудио файлы",
                        "name": "full_audio",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.PointOfInterest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/poi/delete": {
            "delete": {
                "description": "Удаляет точку интереса со всеми связанными файлами",
                "tags": [
                    "POI"
                ],
                "summary": "Удаление точки интереса по id",
                "parameters": [
                    {
                        "type": "number",
                        "example": 55.7558,
                        "description": "Широта",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное выполнение",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/api/poi/nearby": {
            "get": {
                "description": "Возвращает ближайшие точки интереса по координатам",
                "tags": [
                    "POI"
                ],
                "summary": "Поиск ближайшей точки интереса",
                "parameters": [
                    {
                        "type": "number",
                        "example": 55.7558,
                        "description": "Широта",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "example": 37.6173,
                        "description": "Долгота",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 150,
                        "example": 100,
                        "description": "Радиус в метрах",
                        "name": "radius",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PointOfInterest"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Проверяет доступность сервиса",
                "tags": [
                    "Health"
                ],
                "summary": "Проверка здоровья сервиса",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/s3/files/{path}": {
            "get": {
                "description": "Скачивает файл из S3 хранилища по указанному пути",
                "tags": [
                    "S3"
                ],
                "summary": "Получить файл из S3",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"images/photo.jpg\"",
                        "description": "Путь к файлу в S3",
                        "name": "path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл",
                        "schema": {
                            "type": "file"
                        },
                        "headers": {
                            "Content-Length": {
                                "type": "string",
                                "description": "Размер файла в байтах"
                            },
                            "Content-Type": {
                                "type": "string",
                                "description": "MIME-тип файла"
                            }
                        }
                    }
                }
            }
        },
        "/s3/health": {
            "get": {
                "description": "Проверяет соединение с S3 хранилищем",
                "tags": [
                    "S3"
                ],
                "summary": "Проверка соединения с S3 хранилищем",
                "responses": {}
            }
        },
        "/s3/list": {
            "get": {
                "description": "Возвращает список файлов в S3 бакете с возможностью фильтрации по префиксу",
                "tags": [
                    "S3"
                ],
                "summary": "Получить список файлов",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"images/\"",
                        "description": "Префикс для фильтрации файлов",
                        "name": "prefix",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.S3ListResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.File": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_short": {
                    "type": "boolean"
                },
                "mime_type": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "integer"
                }
            }
        },
        "domain.PointOfInterest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "full_audio_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.File"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "image_file": {
                    "$ref": "#/definitions/domain.File"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "short_audio_file": {
                    "$ref": "#/definitions/domain.File"
                }
            }
        },
        "domain.S3FileInfo": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "last_modified": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "domain.S3ListResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.S3FileInfo"
                    }
                }
            }
        },
        "handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        }
    }
}